split shortcut creation into general and os specific parts
